pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azl-bot"
version = "0.1.0"
description = "Azur Lane Bot - Automated game assistant"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    {name = "Coldaine", email = "coldaine@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "PySide6>=6.6.0",
    "paddlepaddle>=2.5.0",
    "paddleocr>=2.7.0",
    "pytesseract>=0.3.10",
    "requests>=2.31.0",
    "uvloop>=0.19.0; sys_platform != 'win32'",
    "sqlalchemy>=2.0.0",
    "rapidfuzz>=3.5.0",
    "loguru>=0.7.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
azl-bot = "azl_bot.ui.app:main"
azl-task = "azl_bot.core.bootstrap:run_task_cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["azl_bot*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true